feat | sprint2 | FRB-173 | messaging 관련 로직 구조 변경 | 조수빈
