+++ src/main/java/com/factoreal/backend/domain/abnormalLog/application/AbnormalLogService.java
@@ -8,10 +8,9 @@ import com.factoreal.backend.domain.abnormalLog.entity.AbnormalLog;
 import com.factoreal.backend.domain.zone.dao.ZoneRepository;
 import com.factoreal.backend.domain.zone.entity.Zone;
 import com.factoreal.backend.domain.abnormalLog.dao.AbnLogRepository;
-import com.factoreal.backend.global.sender.WebSocketSender;
-import com.factoreal.backend.global.kafka.strategy.alarmMessage.RiskMessageProvider;
-import com.factoreal.backend.global.kafka.strategy.enums.RiskLevel;
-import com.factoreal.backend.global.kafka.strategy.enums.SensorType;
+import com.factoreal.backend.messaging.kafka.strategy.alarmMessage.RiskMessageProvider;
+import com.factoreal.backend.messaging.kafka.strategy.enums.RiskLevel;
+import com.factoreal.backend.messaging.kafka.strategy.enums.SensorType;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
