--- src/main/java/com/factoreal/backend/global/kafka/strategy/alarmList/AppPushNotificationStrategy.java
@@ -1,22 +1,0 @@
-package com.factoreal.backend.global.kafka.strategy.alarmList;
-
-import com.factoreal.backend.global.kafka.strategy.enums.AlarmEventDto;
-import com.factoreal.backend.global.kafka.strategy.enums.RiskLevel;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Component;
-
-@Slf4j
-@Component("APP")
-public class AppPushNotificationStrategy implements NotificationStrategy {
-
-    // TODO FCM 전송 로직
-    @Override
-    public void send(AlarmEventDto alarmEventDto) {
-        log.info("📲 App Push Notification Strategy.");
-    }
-
-    @Override
-    public RiskLevel getSupportedLevel() {
-        return RiskLevel.WARNING;
-    }
-}
