--- src/main/java/com/factoreal/backend/global/kafka/strategy/alarmMessage/RiskMessageProvider.java
@@ -1,8 +1,0 @@
-package com.factoreal.backend.global.kafka.strategy.alarmMessage;
-
-import com.factoreal.backend.global.kafka.strategy.enums.RiskLevel;
-import com.factoreal.backend.global.kafka.strategy.enums.SensorType;
-
-public interface RiskMessageProvider {
-    String getMessage(SensorType sensorType, RiskLevel riskLevel);
-}
