--- src/main/java/com/factoreal/backend/global/tmp/ZoneDangerDto.java
@@ -1,16 +1,0 @@
-package com.factoreal.backend.global.tmp;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class ZoneDangerDto {
-    private String zoneId;
-    private String sensorType;
-    private int level;
-}
