+++ src/main/java/com/factoreal/backend/messaging/common/dto/ZoneDangerDto.java
@@ -1,0 +1,16 @@
+package com.factoreal.backend.messaging.common.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class ZoneDangerDto {
+    private String zoneId;
+    private String sensorType;
+    private int level;
+}
