+++ src/main/java/com/factoreal/backend/messaging/kafka/strategy/alarmMessage/RiskMessageProvider.java
@@ -1,0 +1,8 @@
+package com.factoreal.backend.messaging.kafka.strategy.alarmMessage;
+
+import com.factoreal.backend.messaging.kafka.strategy.enums.RiskLevel;
+import com.factoreal.backend.messaging.kafka.strategy.enums.SensorType;
+
+public interface RiskMessageProvider {
+    String getMessage(SensorType sensorType, RiskLevel riskLevel);
+}
