+++ src/main/java/com/factoreal/backend/messaging/kafka/strategy/enums/SensorType.java
@@ -1,0 +1,13 @@
+package com.factoreal.backend.messaging.kafka.strategy.enums;
+
+public enum SensorType {
+    current,
+    dust,
+    temp,
+    humid,
+    vibration,
+    voc;
+    public static SensorType getSensorType(String sensorType){
+        return SensorType.valueOf(sensorType);
+    }
+}
